requirements.txt
Flask
google-generativeai
python-dotenv
reportlab
requests
psycopg2-binary